:art: The style file was created for the pages. Redux structure has been changed.
