+++ src/app/pages/LandingPage.js
@@ -1,18 +1,36 @@
-import React from 'react'
+import React, { useEffect } from 'react'
 import { Footer, NavBar, StyledButton, ProjectCard } from '../components'
 import { Link } from 'react-router-dom'
-import styled from 'styled-components'
 import Example from '../assets/example.svg'
-import { useSelector } from 'react-redux'
+import axios from 'axios'
+import { useDispatch, useSelector } from 'react-redux'
+import { HeaderUp, DarkenBackground, AddSomeMargin, Invite, Center, Description, Container } from './style'
 
-const Container = styled.div`
-justify-content: space-between;
-min-height: 100vh;
-background-color: 'red';
-`
 const LandingPage = () => {
   const { projects } = useSelector((state) => state.projects)
+  const dispatch = useDispatch()
 
+  useEffect(() => {
+    dataFetch()
+  }, [])
+
+  const dataFetch = async () => {
+    try {
+      const query = `{
+        repositories {
+          pathname
+          issues {
+            title
+          }
+        }
+      }`
+      dispatch({ type: 'FETCHING_PROJECTS' })
+      const res = await axios.get(`http://localhost:8080/graphql?query=${query}`)
+      dispatch({ type: 'FETCHED_PROJECTS', payload: res.data.data.repositories })
+    } catch (error) {
+      dispatch({ type: 'ERROR' })
+    }
+  }
   return (
     <Container>
       <NavBar />
@@ -39,11 +57,11 @@ const LandingPage = () => {
         </DarkenBackground>
       </HeaderUp>
       <Center>
-      {projects.map((data, index) => (
+        {projects.map((data, index) => (
           <div key={index}>
             <ProjectCard repository={data.pathname} />
           </div>
-      ))}
+        ))}
       </Center>
       <div style={{ marginRight: 480, marginLeft: 480 }}>
         <div style={{ display: 'flex', alignItems: 'flex-start', justifyContent: 'flex-start' }}>
@@ -65,68 +83,5 @@ const LandingPage = () => {
     </Container >
   )
 }
-
-const HeaderUp = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  @media screen and (min-width: 1024px) {
-    height: 50%;
-  }
-
-  @media screen and (min-width: 768px) {
-    height: 50%;
-  }
-
-  @media screen and (min-width: 320px) {
-    height: 50%;
-  }
-  background: url('https://image.freepik.com/free-vector/mars-landscape-alien-planet-martian-background_107791-1781.jpg') no-repeat center/cover;
-  text-align: center;
-`
-
-const Invite = styled.div`
-  margin-top:  1em;
-`
-const Description = styled.div`
-  width: 612px;
-  height: 200px;
-  padding: 10px;
-  display: flex;
-  justify-content: center;
-
-`
-
-const Center = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-top: 1em;
-  flex-direction: row;
-  @media screen and (min-width: 1024px) {
-    width: 100%;
-  }
-
-  @media screen and (min-width: 768px) {
-    width: 100%;
-  }
-
-  @media screen and (min-width: 320px) {
-    width: 100%;
-  }
-`
-
-const DarkenBackground = styled.div`
-  width: 100%;
-  height: 100%;
-`
-
-const AddSomeMargin = styled.div`
-  margin: 5em;
-  padding: 25px;
-  background: rgba(255, 255, 255, .15);
-  border-radius: 5px;
-`
 
 export default LandingPage
